feat(build): add support for windows filesystem
When building on the Windows filesystem back-slashes were being used for
file paths and module ids.  The back-slashes were being escaped but in
the case of modules that start with the letter 'u' ended up being
interpreted as starting a unicode character.  To fix this we are
replacing back-slashes with forward-slashes in module paths and module
ids.

This is believed to be a non-breaking change.